Create nbody6.f90
Most of the computation time in the nbody benchmark is spent in subroutine advance(). I have rearranged the three sets of DO loops in this subroutine into two sets. On my PC, with Intel OneAPI 2021.5, I gained a speed increase of 18 percent as a result of this rearrangement.

In this version, the special technique of using the X64 SSE instruction RSQRTSS to generate a single precision approximation to 1/sqrt(x), followed by a single iteration of double precision Newton-Raphson polishing the root, which is used in the Julia version of Nbody, is not used since it did not yield better performance than the straightforward calculation of x^(-3/2), for a given x.